Add GeoMath class with geographic coordinates functions and tests
Implemented a GeoMath class with static methods for calculating distances between two points on the Earth's surface, checking proximity to line segments, verifying if a point is on a polyline, finding the next point on a route, and calculating the distance to the next point. Also included tests for these functionalities. Additionally, introduced a method to obtain the bounding coordinates of a set of routes.

Note: Some methods are marked as "in develop" and may not be fully functional yet.